refactor: share supporter image helper
